Changed the Stopping Counter to ask for seconds and not periodic loops.

+++ src/main/java/entech/util/PeriodicLoopsPerSecond.java
@@ -2,9 +2,9 @@ package entech.util;
 
 public class PeriodicLoopsPerSecond {
 
-  private static int PERIODIC_LOOP_FACTOR_SEC = 50;
+  private static double PERIODIC_LOOP_FACTOR_SEC = 50;
 
-  public static int getLoopsPerSecond(int seconds) {
+  public static double getLoopsPerSecond(double seconds) {
     return PERIODIC_LOOP_FACTOR_SEC * seconds;
   }
 }
